+++ builder/push.go
@@ -11,7 +11,7 @@ import (
 )
 
 const (
-	maxPushRetries = 5
+	maxPushRetries = 3
 )
 
 func (b *Builder) pushWithRetries(ctx context.Context, images []string) error {
